% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{AddDerivedVariables}
\alias{AddDerivedVariables}
\title{Add derived design variables to a subject table}
\usage{
AddDerivedVariables(subjectTable, configuration)
}
\arguments{
\item{subjectTable}{A subject-level table.}

\item{configuration}{A networkR configuration object.}
}
\value{
The subject table with derived variables added.
}
\description{
Add derived design variables to a subject table
}
