% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\title{Transform network data into an adjacency matrix}
\usage{
adjacencyMatrix(edgeList, nodeList)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}
}
\description{
This function transforms network data into an n-by-n adjacency matrix.
}
\examples{
adjacencyMatrix()
}
\keyword{adjacency}
\keyword{matrix}

