% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliationEnvironment.R
\name{affiliationEnviornment}
\alias{affiliationEnviornment}
\title{Find the affiliation neighbourhood of each affiliation}
\usage{
affiliationEnviornment(edgeList, affiliationList)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{affiliationList}{A dataframe where all affiliations and their respective names are listed.}
}
\description{
This function transforms network data into an n-by-n adjacency matrix.
}
\examples{
affiliationEnviornment()
}
\keyword{affiliation}
\keyword{environment}

