% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliationInfluence.R
\name{affiliationInfluence}
\alias{affiliationInfluence}
\title{Calculate the influence of affiliations within a hypergraph}
\usage{
affiliationInfluence(edgeList, affiliationList, weights, normalise)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{affiliationList}{A dataframe where all affiliations and their respective names are listed.}

\item{normalise}{Should the influence score be normalised? TRUE or FALSE.}

\item{weight}{The weight or value of an affiliation.}
}
\description{
The function calculates the influence of nodes and affiliations within a hypergraph. The measure accepts weights, or values, distributed over affiliations.
}
\examples{
affiliationInfluence()
}
\keyword{influence}
