% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliationProjection.R
\name{affiliationProjection}
\alias{affiliationProjection}
\title{Represent a hypergraph as a network of affiliations}
\usage{
affiliationProjection(hypergraph)
}
\arguments{
\item{hypergraph}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}
}
\description{
This function projects a hypergraph into a network of affiliations.
}
\examples{
affiliationProjection()
}
\keyword{affiliation}
\keyword{projection}

