% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_network.R
\name{attach_attributes}
\alias{attach_attributes}
\title{Attach Attributes to Network Nodes}
\usage{
attach_attributes(network, attributes)
}
\arguments{
\item{network}{A \code{network} object created by \code{graph_to_network}.}

\item{attributes}{A \code{data.frame} where each row represents attributes for a node in \code{network}.
The number of rows in \code{attributes} must equal the number of unique nodes in \code{network}.}
}
\value{
A modified \code{network} object with the \code{attributes} field populated by the input \code{data.frame}.
}
\description{
Adds a \code{data.frame} of attributes to the nodes in a \code{network} object. Each row in \code{attributes} corresponds to a node,
with row names matching the node indices in the network. The function verifies the correct number of attributes
to ensure each node has one associated row of attributes.
}
\examples{
graph <- list(c(1, 2), c(2, 3), c(3, 1))
network <- graph_to_network(graph)
attributes <- data.frame(type = c("A", "B", "C"), row.names = c(1, 2, 3))
network <- attach_attributes(network, attributes)
print(network$attributes)

}
