% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockSNE.R
\name{blockSNE}
\alias{blockSNE}
\title{Blocks formed in a Strong Nash Equilibrium}
\usage{
blockSNE(network, nodeNames, c, s, adjMatrix, setPS, setPower, approximate)
}
\arguments{
\item{network}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe within which all nodes and their respective names are listed.}

\item{c}{The cost of forming a connection to a single other node in order to form a block.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{setPower}{The brokerage power of each set of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function finds all sets of nodes that form blocks in a Strong Nash Equilibrium along with their respective brokerage power.
}
\examples{
blockSNE()
}
\keyword{Equilibrium}
\keyword{Nash}
\keyword{String}
\keyword{blocks}

