% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\name{BuildBlacklist}
\alias{BuildBlacklist}
\title{Build a blacklist from variable tiers and study design}
\usage{
BuildBlacklist(discretizationResult, configuration)
}
\arguments{
\item{discretizationResult}{A result from \code{DiscretizeMeasurements()}.}

\item{configuration}{A networkR configuration object.}
}
\value{
A list containing the blacklist and active exogenous variables.
}
\description{
Build a blacklist from variable tiers and study design
}
