% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{BuildConfiguration}
\alias{BuildConfiguration}
\title{Build a networkR configuration object}
\usage{
BuildConfiguration(overrides = list())
}
\arguments{
\item{overrides}{A named list of configuration overrides.}
}
\value{
A validated configuration object.
}
\description{
Build a networkR configuration object
}
