% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_families}
\alias{cluster_families}
\title{Cluster families}
\usage{
cluster_families(id, fid, mid)
}
\arguments{
\item{id}{Numeric vector of ids}

\item{fid}{Numeric vector of ids of the father}

\item{mid}{Numeric vector of ids of the mother}
}
\value{
Returns an integer vector giving the family index
}
\description{
Computes a vector of groupings in families based on id, father id, and mother id. No check is done to ensure that the id, fid, and mid actually refere to a proper family structure. References to ids in the fid and mid arguments that are not part of the id vector are considered founders.
}
\examples{

id <- 1:11
fid <- c(NA, NA, 1, 1, NA, 23, 45, 5, 5, 7, NA)
mid <- c(NA, NA, 2, 2, 65, NA, 46, 6, 6, 6, NA)
cluster_families(id, fid, mid)

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{manip}
