% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_network}
\alias{cluster_network}
\title{Cluster network}
\usage{
cluster_network(from, to, weight, threshold = 0)
}
\arguments{
\item{from}{Integer vector of nodes where the link is from}

\item{to}{Integer vector of nodes where the link is to}

\item{weight}{Numeric vector of link weights}

\item{threshold}{A double indicating the minimum link weight that should be considered. Only weights equal to or greater than the threshold will be considered in the clustering}
}
\value{
Returns an integer vector giving the cluster index for each node
}
\description{
Identifies cluster in a symmetric, weighted network based on an edgelist. The edgelist should contain positive integers for the node numbers. Sequential node numbers will speed up computations.
}
\examples{

aaa <- structure(c(1, 6, 5, 4, 5, 11, 1, 5, 3, 13, 16, 15, 18,
                   6, 7, 8, 10, 12, 14, 15, 15, 16, 17, 17, 18, 20), .Dim = c(13L, 2L))
cluster_network(aaa[,1], aaa[,2], weight=rep(1, nrow(aaa))) 

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{manip}
