% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{connectivity}
\alias{connectivity}
\title{Connectivity of a network}
\usage{
connectivity(network, nodeNames, adjMatrix)
}
\arguments{
\item{network}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe within which all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
This function calculates the connectivity of a network by summing the number of unique successors of each node within the network.
}
\examples{
connectivity()
}
\keyword{connectivity}

