% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalCoverage.R
\name{criticalCoverage}
\alias{criticalCoverage}
\title{Per capita coverage of all critical sets}
\usage{
criticalCoverage(edgeList, nodeList, s, setPS, approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates per capita coverage of all critical sets in the network.
}
\examples{
criticalCoverage()
}
\keyword{coverage}
