% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalSets.R
\name{criticalSets}
\alias{criticalSets}
\title{Find all critical sets in a network}
\usage{
criticalSets(edgeList, nodeList, s, adjMatrix, setPS, setPower, approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{setPower}{All sets within a network along with their respective brokerage scores.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates all critical sets in the network.
}
\examples{
criticalSets()
}
\keyword{critical}
\keyword{sets}

