% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeDistribution.R
\name{degreeDistribution}
\alias{degreeDistribution}
\title{Plot degree distribution}
\usage{
degreeDistribution(network, nodeNames)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed.}
}
\description{
This function plots the degree distribution of a network. Depends on `ggplot2`.
}
\examples{
degreeDistribution()
}
\keyword{degree}
\keyword{distribution}

