% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density}
\alias{density}
\title{Density of a network}
\usage{
density(edgeList, nodeList, undirected)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{undirected}{Is the network graph undirected? TRUE or FALSE.}
}
\description{
This function calculates the density of a network graph.
}
\examples{
density()
}
\keyword{density}

