% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{DiscretizeMeasurements}
\alias{DiscretizeMeasurements}
\title{Discretize continuous measurements for Bayesian network learning}
\usage{
DiscretizeMeasurements(imputationResult, configuration)
}
\arguments{
\item{imputationResult}{An imputation result created by \code{ImputeMeasurements()}.}

\item{configuration}{A networkR configuration object.}
}
\value{
A list containing continuous and discrete analysis data.
}
\description{
Discretize continuous measurements for Bayesian network learning
}
