% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elites.R
\name{elites}
\alias{elites}
\title{Find all elites in the hypergraph}
\usage{
elites(edgeList, nodeList, affiliationList, typeList, e)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{nodeList}{A dataframe where all node numbers and their respective names are listed.}

\item{affiliationList}{A dataframe where all affiliation numbers and their respective names are listed.}

\item{typeList}{Each affiliation is assigned a type; must be in order of affiliations.}

\item{e}{The number of aspects that a node exists in order for it to be determined as an elite.}
}
\description{
This function lists all elites nodes in a hypergraph determined by some criteria.
}
\examples{
elites()
}
\keyword{elite}
