% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalInfluence.R
\name{externalInfluence}
\alias{externalInfluence}
\title{Calculate the external influence of affiliations}
\usage{
externalInfluence(edgeList, affiliationList, weight)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{affiliationList}{A dataframe where all affiliations and their respective names are listed.}

\item{weight}{The weight or value of an affiliation.}
}
\description{
This function calculates the external influence of affiliations within a hypergraph. Affiliations can have values, and therefore be weighted.
}
\examples{
externalInfluence()
}
\keyword{influence}

