% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{hits}
\alias{hits}
\title{Create adjacency matrix}
\usage{
hits(adjmatrix, maxiter = 25L)
}
\arguments{
\item{adjmatrix}{an adjacency matrix}

\item{maxiter}{non-negative integer}
}
\value{
Returns a sparse adjacency matrix
}
\description{
Split a matrix into block diagonal sub matrices according to clusters and
combine the lower triangular parts into a vector
}
\examples{

from <- c("A", "A", "A", "B", "C")
to <- c("B", "C", "D", "D", "E")
adjacency(from, to) \%>\% hits()

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\keyword{manip}
