% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergraphNeighbourhood.R
\name{hypergraphNeighbourhood}
\alias{hypergraphNeighbourhood}
\title{Find the neighbourhood of each node}
\usage{
hypergraphNeighbourhood(edgeList, nodeList)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}
}
\description{
This function finds the neighbourhood of each node within a given hypergraph.
}
\examples{
hypergraphNeighbourhood()
}
\keyword{neighbour}
\keyword{neighbourhood}

