% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ImputeMeasurements}
\alias{ImputeMeasurements}
\title{Impute missing measurement values with MICE}
\usage{
ImputeMeasurements(subjectTable, configuration)
}
\arguments{
\item{subjectTable}{A subject-level table.}

\item{configuration}{A networkR configuration object.}
}
\value{
A list containing imputed data and analysis metadata.
}
\description{
Impute missing measurement values with MICE
}
