% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUndirected.R
\name{isUndirected}
\alias{isUndirected}
\title{Query whether a network is undirected}
\usage{
isUndirected(edgeList, nodeList, adjMatrix)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that queries whether some network data is undirected and returns a TRUE or FALSE value.
}
\examples{
isUndirected()
}
\keyword{undirected}
