% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeCoverage.R
\name{nodeCoverage}
\alias{nodeCoverage}
\title{Calculate the coverage of each node}
\usage{
nodeCoverage(edgeList, nodeList, s, adjMatrix, setPS, approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates the coverage for each node based on their existence within blocks.
}
\examples{
nodeCoverage()
}
\keyword{coverage}
\keyword{node}

