% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{nodeInfluence}
\alias{nodeInfluence}
\title{Calculate the  influence of nodes and affiliations}
\usage{
nodeInfluence(hypergraph, nodeNames, affiliationNames, weights)
}
\arguments{
\item{hypergraph}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed.}

\item{affiliationNames}{A dataframe where all affiliations and their respective names are listed.}

\item{weight}{The weight or value of an affiliation.}
}
\description{
The function calculates the influence of nodes and affiliations within a hypergraph. The measure accepts weights, or values, distributed over affiliations.
}
\examples{
nodeInfluence()
}
\keyword{influence}

