% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeProjection.R
\name{nodeProjection}
\alias{nodeProjection}
\title{Represent a hypergraph as a network of nodes only}
\usage{
nodeProjection(edgeList)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}
}
\description{
This function projects a hypergraph into a network of nodes only.
}
\examples{
nodeProjection
}
\keyword{node}
\keyword{projection}

