% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSetBrokerage.R
\name{nodeSetBrokerage}
\alias{nodeSetBrokerage}
\title{Brokerage of a node given its presence in a node set}
\usage{
nodeSetBrokerage(edgeList, nodeList, s, perCapita, adjMatrix, setPS, setPower,
  approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{perCapita}{The brokerage power of the node set divided by the number of nodes within the set.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates the brokerage of each node given its existence in critical sets.
}
\examples{
nodeSetBrokerage()
}
\keyword{brokerage}
