% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ParseConfiguration}
\alias{ParseConfiguration}
\title{Parse a YAML configuration file}
\usage{
ParseConfiguration(filePath)
}
\arguments{
\item{filePath}{Path to a YAML configuration file.}
}
\value{
A validated configuration object.
}
\description{
Parse a YAML configuration file
}
