% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentialBrokerage.R
\name{potentialBrokerage}
\alias{potentialBrokerage}
\title{Potential brokerage of a network}
\usage{
potentialBrokerage(edgeList, nodeList, adjMatrix)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that calculates the number of indirect relationships that exist, and therefore can be potentially brokered. This acts as a normaliser for `middlemanPower`.
}
\examples{
potentialBrokerage()
}
\keyword{brokerage}
\keyword{middleman}

