% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_to_network.R
\name{print.network}
\alias{print.network}
\title{Print method for Network Object}
\usage{
\method{print}{network}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{network}.}

\item{digits}{Integer specifying the number of significant digits to use
for displaying the mean degree. Defaults to the maximum of 3 or the current
option set by \code{getOption("digits") - 3}.}

\item{...}{Additional arguments passed to other methods (not used here).}
}
\value{
Invisibly returns \code{x}.
}
\description{
This is a print method for objects of class \code{network}, displaying basic
summary information about the network, including the number of nodes, edges,
and mean degree.
}
\examples{
# Create a network object and print summary
graph <- list(c(1, 2), c(2, 3), c(3, 1))
network <- graph_to_network(graph)
print(network)

}
