% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGraph.R
\name{randomGraph}
\alias{randomGraph}
\title{Generate a random network}
\usage{
randomGraph(n, p, undirected)
}
\arguments{
\item{n}{The number of nodes to be present within the network.}

\item{p}{The probability that any two nodes are connected together.}

\item{undirected}{should the randomly generated network be undirected? TRUE or FALSE.}
}
\description{
The function generates a randomly connected network.
}
\examples{
randomGraph()
}
\keyword{graph}
\keyword{network}
\keyword{random}
