% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ReadPanels}
\alias{ReadPanels}
\title{Read panel files into a joined subject table}
\usage{
ReadPanels(configuration)
}
\arguments{
\item{configuration}{A networkR configuration object.}
}
\value{
A joined subject-level table.
}
\description{
Read panel files into a joined subject table
}
