% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{robustness}
\alias{robustness}
\title{Calculate the robustness of a middleman position}
\usage{
robustness(edgeList, nodeList)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}
}
\description{
Robustness of a middleman measured by the number of arcs that need to be removed in order for a middleman to lose its position as a middleman.
}
\examples{
robustness()
}
\keyword{robustness}
