% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCoverage.R
\name{setCoverage}
\alias{setCoverage}
\title{Calculate the coverage of all sets of nodes}
\usage{
setCoverage(edgeList, nodeList, s, adjMatrix, setPS, approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates the coverage, i.e., the number of indirect connections that the node set faciliates, for each set of nodes in the network.
}
\examples{
setCoverage()
}
\keyword{coverage}

