% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCriticality.R
\name{setCriticality}
\alias{setCriticality}
\title{Calculate the criticality of all node sets in Strong Nash equilibrium (SNE)}
\usage{
setCriticality(network, nodeNames, c, s, adjMatrix, setPS, setPower,
  approximate)
}
\arguments{
\item{network}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe within which all nodes and their respective names are listed.}

\item{c}{The cost of signalling to, and adding an, extra node to a block.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{setPS}{The set of predeccessors and successors for each combination of nodes considered.}

\item{approximate}{Should the Strong Nash Equilibrium be approximated? TRUE or FALSE.}
}
\description{
This function calculates the criticality of all node sets that remain in all resulting SNE configurations of a block formation game.
}
\examples{
setCriticality()
}
\keyword{criticality}

