% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPredSucc.R
\name{setPredSucc}
\alias{setPredSucc}
\title{Find the predecessors and successors of a node set}
\usage{
setPredSucc(edgeList, nodeList, s, adjMatrix, approximate)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{s}{The maximum size of block that is considered within the block formation game.}

\item{adjMatrix}{The network represented as an adjacency matrix.}

\item{approximate}{Should the number of sets considered be truncated? TRUE or FALSE.}
}
\description{
This function calculates the predecessors and successors of a node or node set.
}
\examples{
setPredSucc()
}
\keyword{node}
\keyword{predecessors}
\keyword{successors}

