% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{simple_impute_measurements}
\alias{simple_impute_measurements}
\title{Impute missing measurement values with MICE}
\usage{
simple_impute_measurements(analysis_data, measurement_columns)
}
\arguments{
\item{subjectTable}{A subject-level table.}

\item{configuration}{A networkR configuration object.}
}
\value{
A list containing imputed data and analysis metadata.
}
\description{
Impute missing measurement values with MICE
}
