% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strongWeak.R
\name{strongWeak}
\alias{strongWeak}
\title{Distinguish strong, weak, and Non-middlemen}
\usage{
strongWeak(edgeList, nodeList, adjMatrix)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that returns whether individual nodes are strong, weak or non-middlemen
}
\examples{
strongWeak()
}
\keyword{brokerage}
\keyword{middleman}
