% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ValidateConfiguration}
\alias{ValidateConfiguration}
\title{Validate a networkR configuration object}
\usage{
ValidateConfiguration(configuration)
}
\arguments{
\item{configuration}{A configuration list.}
}
\value{
The validated configuration object, invisibly.
}
\description{
Validate a networkR configuration object
}
