% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Create adjacency matrix}
\usage{
adjacency(from, to, type = 1, directed = TRUE)
}
\arguments{
\item{from}{a vector}

\item{to}{a vector}

\item{type}{numeric}

\item{directed}{logical. Are the edges directed (TRUE, the default) or bidirected(FALSE).}
}
\value{
Returns a sparse adjacency matrix
}
\description{
Split a matrix into block diagonal sub matrices according to clusters and
combine the lower triangular parts into a vector
}
\examples{

from <- 1:5

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\keyword{manip}
