% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyDF.R
\name{adjacencyDF}
\alias{adjacencyDF}
\title{Transform network data into a data frame adjacency matrix}
\usage{
adjacencyDF(network, nodeNames)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed}
}
\description{
This function transforms network data into an n-by-n adjacency matrix dressed up as a data frame
}
\examples{
adjacencyDF()
}
\keyword{adjacency}

