% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_families}
\alias{cluster_families}
\title{Cluster families}
\usage{
cluster_families(id, fid, mid)
}
\arguments{
\item{id}{Numeric vector of ids}

\item{fid}{Numeric vector of ids of the father}

\item{mid}{Numeric vector of ids of the father}
}
\value{
Returns a vector giving the xxxxx
zero for equal correlation matrices and unity if they differ to a maximum extent.
}
\description{
Computes the correlation matrix distance between two correlation matrices
}
\examples{

x <- 1:5

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}
