% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_families}
\alias{cluster_families}
\title{Cluster families}
\usage{
cluster_families(id, fid, mid)
}
\arguments{
\item{id}{Numeric vector of ids}

\item{fid}{Numeric vector of ids of the father}

\item{mid}{Numeric vector of ids of the mother}
}
\value{
Returns an integer vector giving the family index
}
\description{
Computes a vector giving the grouping in families based on id, father id, and mother id
}
\examples{

x <- 1:5

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{manip}
