% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_network}
\alias{cluster_network}
\title{Cluster network}
\usage{
cluster_network(from, to, weight, threshold = 0)
}
\arguments{
\item{from}{Integer vector of nodes where the link is from}

\item{to}{Integer vector of nodes where the link is to}

\item{weight}{Numeric vector of link weights}

\item{threshold}{A double indicating the minimum link weight that should be considered. Only weights equal to or greater than the threshold will be considered in the clustering}
}
\value{
Returns an integer vector giving the cluster index for each node
}
\description{
Identifies cluster in a symmetric, weighted network based on an edgelist. The edgelist should contain positive integers for the node numbers. Sequential node numbers will speed up computations.
}
\examples{

cluster_network(from, to, weight)

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{manip}
