% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Calculate node degree}
\usage{
degree(network, nodeNames)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed}
}
\description{
This function calculates the degree of all nodes in the network. This includes counting the in-degree and out-degree of each node.
}
\examples{
degree()
}
\keyword{degree}

