% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNetwork.R
\name{filterNetwork}
\alias{filterNetwork}
\title{Filter network}
\usage{
filterNetwork(network)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column.}
}
\description{
This function filters a network such that it removes any duplicated connections, or more specifically arcs, between a pair of nodes.
}
\examples{
filterNetwork()
}
\keyword{filter}

