% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{hits}
\alias{hits}
\title{Hyperlink-induced topic search}
\usage{
hits(adjmatrix, maxiter = 25L)
}
\arguments{
\item{adjmatrix}{an adjacency matrix}

\item{maxiter}{non-negative integer}
}
\value{
Returns a list of authorities and hubs
}
\description{
Hyperlink-induced topic search (HITS) is a link analysis algorithm
that is also known as hubs and authorities. It rates nodes by
comparing arrows pointing in and out of nodes in an assymetrical graph.
}
\examples{

from <- c("A", "A", "A", "B", "C")
to <- c("B", "C", "D", "D", "E")
hits(adjacency(from, to))

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\keyword{manip}
