% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middlemanPower.R
\name{middlemanPower}
\alias{middlemanPower}
\title{Calculate Middleman power for all nodes}
\usage{
middlemanPower(edgeList, nodeList, adjMatrix, normalised)
}
\arguments{
\item{edgeList}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that calculates the middleman power of each node in a network.
}
\examples{
middlemanPower()
}
\keyword{brokerage}
\keyword{middleman}
