% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeInfluence.R
\name{nodeInfluence}
\alias{nodeInfluence}
\title{Calculate the influence of nodes within a hypergraph}
\usage{
nodeInfluence(edgeList, nodeList, affiliationList, weights, normalise)
}
\arguments{
\item{edgeList}{A dataframe of network data where nodes are in the first column and affiliations are in the second column. Nodes are members of the affiliation that they are next to.}

\item{nodeList}{A dataframe where all nodes and their respective names are listed.}

\item{affiliationList}{A dataframe where all affiliations and their respective names are listed.}

\item{normalise}{Should the influence score be normalised? TRUE or FALSE.}

\item{weight}{The weight or value of an affiliation.}
}
\description{
The function calculates the influence of nodes and affiliations within a hypergraph. The measure accepts weights, or values, distributed over affiliations.
}
\examples{
nodeInfluence()
}
\keyword{influence}
