% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predecessorsSuccessors.R
\name{predecessorsSuccessors}
\alias{predecessorsSuccessors}
\title{Find the predecessors and successors of a node}
\usage{
predecessorsSuccessors(edgeList, nodeList, adjMatrix)
}
\arguments{
\item{edgeList}{A dataframe of network data within which sources are in the first column and targets are in the second column.}

\item{nodeList}{A dataframe within which all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
This function calculates the predecessors and successors of a node or node set.
}
\examples{
predecessorsSuccessors()
}
\keyword{node}
\keyword{predecessors}
\keyword{successors}
