% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middlemanPowerDetail.R
\name{strongWeak}
\alias{strongWeak}
\title{Provide detailed results of middleman power}
\usage{
strongWeak(network, nodeNames, adjMatrix)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that returns detailed reults on each nodes middleman power in the network.
}
\examples{
middlemanPowerDetail()
}
\keyword{brokerage}
\keyword{middleman}

