% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirectedAdjMatrix.R
\name{undirectedAdjMatrix}
\alias{undirectedAdjMatrix}
\title{Coerce adjacency matrix to be undirected}
\usage{
undirectedAdjMatrix(network, nodeNames, adjMatrix)
}
\arguments{
\item{network}{A dataframe of network data where sources are in the first column and targets are in the second column.}

\item{nodeNames}{A dataframe where all nodes and their respective names are listed.}

\item{adjMatrix}{The network represented as an adjacency matrix.}
}
\description{
A function that coerces an adjacency matrix to become undirected.
}
\examples{
undirectedAdjMatrix()
}
\keyword{undirected}

